Update condition to work only for mac

+++ src/main/java/com/appium/manager/AppiumDriverManager.java
@@ -161,14 +161,18 @@ public class AppiumDriverManager {
     }
 
     public void stopAppiumDriver() throws Exception {
+        String OS = System.getProperty("os.name").toLowerCase();
+        String command;
         if (AppiumDeviceManager.getAppiumDevice().getDevice().getUdid().length()
                 == IOSDeviceConfiguration.IOS_UDID_LENGTH) {
             String hostName = AppiumDeviceManager.getAppiumDevice().getHostName();
             AppiumManagerFactory.getAppiumManager(hostName).destoryIOSWebKitProxy(hostName);
         }
         if(AppiumDeviceManager.getAppiumDevice().getChromeDriverPort() > 0) {
-            String command = "kill -9 $(lsof -ti tcp:" + AppiumDeviceManager.getAppiumDevice().getChromeDriverPort() + ")";
-            new CommandPrompt().runCommand(command);
+            if(OS.indexOf("mac") >= 0) {
+                command = "kill -9 $(lsof -ti tcp:" + AppiumDeviceManager.getAppiumDevice().getChromeDriverPort() + ")";
+                new CommandPrompt().runCommand(command);
+            }
             AppiumDeviceManager.getAppiumDevice().setChromeDriverPort(0);
         }
         if (AppiumDriverManager.getDriver() != null
